+++ lib/presentation/screens/redeem_points_screen.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import '../utils/constant.dart';
+import 'package:loyalty_program_frontend/presentation/utils/helpers/size_helper.dart';
+
+
 
 class RedeemPointsScreen extends StatefulWidget {
   BoxConstraints boxConstraints;
@@ -139,7 +141,7 @@ class _RedeemPointsScreenState extends State<RedeemPointsScreen> {
                                   ),
                                   const SizedBox(width: 10),
                                   Image.asset(
-                                    'packages/loyalty_program_frontend/assets/reward_point_coin.png',
+                                    'packages/loyalty_program_frontend/assets/images/reward_point_coin.png',
                                     width: size(widget.boxConstraints, 24),
                                     height: size(widget.boxConstraints, 24),
                                   )
