+++ lib/presentation/utils/helpers/loyalty_points_checker.dart
